Moved bibtex fuction to util
