+++ backend/services/references/index.js
@@ -1,0 +1,1 @@
+export { default } from "./routes/index.js";
